structure: {controller, service, repository,..} -> {domain-{article, member, channel,..}}
