+++ src/main/java/com/chatter/Chatly/domain/common/FriendState.java
