+++ src/main/java/com/chatter/Chatly/domain/entity/ChatRoom.java
