+++ src/main/java/com/chatter/Chatly/domain/entity/Friend.java
