+++ src/main/java/com/chatter/Chatly/domain/entity/Ownable.java
@@ -1,0 +1,6 @@
+package com.chatter.Chatly.entity;
+
+// Member를 가진 객체 - MemberId 반환
+public interface Ownable<T> {
+    T getOwnerId();
+}
