+++ src/main/java/com/chatter/Chatly/domain/entity/Privilege.java
